Refactored Invoice entity
