--- domain/src/invoices/entities/types.d.ts
@@ -1,11 +1,0 @@
-interface InvoiceExpenseDto {
-  expense: string
-  days: number
-  value: number
-}
-
-interface InvoiceGuestDto {
-  name: string
-  total: number
-  expenses: InvoiceExpenseDto[]
-}
